requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
PyMySQL==1.1.0
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.0
urllib3==2.4.0
